Refactor navigation tests
